requirements.txt
discord.py~=2.4.0
